We need to compensate for the time it takes to read the sensor.
If the climb rate is 20cm in 0.5 seconds, then over 1 second the climb rate is 40cm. If the climb rate 20cm in 2 seconds, then over 1 second the climb rate is 10cm. I'm not entirely sure the code is right but it should match the method above. The climb rate calculations need more testing that I'm not doing right now.